testing, adding validation to update
