+++ apps/backend/package.json
@@ -51,7 +51,8 @@
     "sequelize": "^6.20.1",
     "sequelize-typescript": "^2.1.3",
     "tiny-invariant": "^1.2.0",
-    "zod": "^3.17.3"
+    "zod": "^3.17.3",
+    "umzug": "^3.1.1"
   },
   "devDependencies": {
     "@swc-node/register": "^1.5.1",
